# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IrForge - Telegram Bot Builder Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/telegram:
    post:
      operationId: linkTelegram
      tags: [auth]
      summary: Link the current account to a Telegram identity via the Login Widget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramAuthInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Missing required fields
        "401":
          description: Invalid Telegram hash, or not logged in
        "409":
          description: Telegram account already linked to a different user
        "503":
          description: Telegram login not configured (TELEGRAM_BOT_TOKEN missing)

  # Users
...
  /marketplace/items:
    get:
      operationId: listMarketplaceItems
      tags: [marketplace]
      summary: List marketplace items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceItem"
  /marketplace/items/{itemId}:
    get:
      operationId: getMarketplaceItem
      tags: [marketplace]
      summary: Get a marketplace item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceItem"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, plan, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        plan:
          type: string
          enum: [free, starter, pro, enterprise]
        botCount:
          type: integer
        telegramId:
          type: ["string", "null"]
        telegramUsername:
          type: ["string", "null"]
        telegramFirstName:
          type: ["string", "null"]
        telegramLastName:
          type: ["string", "null"]
        telegramPhotoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    TelegramAuthInput:
      type: object
      description: >
        Raw payload returned by Telegram's official Login Widget
        (https://core.telegram.org/widgets/login). Verified server-side
        against TELEGRAM_BOT_TOKEN before being linked to the current user.
        Telegram never includes a phone number in this payload.
      required: [id, first_name, auth_date, hash]
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        username:
          type: string
        photo_url:
          type: string
        auth_date:
          type: integer
        hash:
          type: string
...
    MarketplaceItem:
      type: object
      required: [id, name, description, category, price, author, version, rating, installCount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [plugin, component, template, theme]
        price:
          type: number
        isFree:
          type: boolean
        author:
          type: string
        version:
          type: string
        rating:
          type: number
        installCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        icon:
          type: ["string", "null"]
        featured:
          type: boolean
        createdAt:
          type: string

    # Plans
...